refactor: улучшает модули модального окна и комментариев после видеовстречи с наставником
