+++ index.html
@@ -208,14 +208,6 @@
     </a>
   </template>
 
-  <!-- Шаблон комментария к фотографии -->
-  <template id="comment">
-    <li class="social__comment">
-      <img class="social__picture" src="" alt="" width="35" height="35">
-      <p class="social__text"></p>
-    </li>
-  </template>
-
   <!-- Сообщение с ошибкой загрузки изображения -->
   <template id="error">
     <section class="error">
