+++ js/gallery/comments.js
@@ -1,0 +1,23 @@
+const container = document.querySelector('.social__comments');
+const template = container.querySelector('.social__comment');
+const shownCounter = document.querySelector('.social__comment-shown-count');
+const totalCounter = document.querySelector('.social__comment-total-count');
+const loaderButton = document.querySelector('.comments-loader');
+
+const createComments = (commentsData) => commentsData.map((properties) => {
+  const {avatar, message, name} = properties;
+  const comment = template.cloneNode(true);
+  comment.querySelector('.social__picture').src = avatar;
+  comment.querySelector('.social__picture').alt = name;
+  comment.querySelector('.social__text').textContent = message;
+  return comment;
+});
+
+const renderComments = (commentsData) => {
+  container.replaceChildren(...createComments(commentsData));
+  shownCounter.textContent = commentsData.length;
+  totalCounter.textContent = commentsData.length;
+  loaderButton.classList.add('hidden');
+};
+
+export {renderComments};
