+++ js/gallery/main.js
@@ -1,9 +1,11 @@
+import {renderThumbnails} from './thumbnails.js';
 import {renderModal} from './modal.js';
 
-function initGallery() {
-  document.addEventListener('thumbnailSelect', (evt) => {
-    renderModal(evt.detail);
+function renderGallery(picturesData) {
+  renderThumbnails(picturesData);
+  document.addEventListener('thumbnailSelect', (event) => {
+    renderModal(event.detail);
   });
 }
 
-export {initGallery};
+export {renderGallery};
