+++ js/main.js
@@ -1,6 +1,4 @@
 import {createPicturesData} from './data.js';
-import {renderThumbnails} from './gallery/thumbnails.js';
-import {initGallery} from './gallery/main.js';
+import {renderGallery} from './gallery/main.js';
 
-renderThumbnails(createPicturesData());
-initGallery();
+renderGallery(createPicturesData());
